Moving Vehicle component to common folder
