+++ src/client/components/common/Vehicle/Vehicle.js
