+++ src/client/components/common/Vehicle/index.js
